Remove domains_v2 datasources with docs
