+++ website/selectel.erb
@@ -16,12 +16,6 @@
             <li<%= sidebar_current("docs-selectel-datasource-domains-domain-v1") %>>
               <a href="/docs/providers/selectel/d/domains_domain_v1.html">selectel_domains_domain_v1</a>
             </li>
-            <li<%= sidebar_current("docs-selectel-datasource-domains-zone-v2") %>>
-              <a href="/docs/providers/selectel/d/domains_zone_v2.html">selectel_domains_zone_v2</a>
-            </li>
-            <li<%= sidebar_current("docs-selectel-datasource-domains-rrset-v2") %>>
-              <a href="/docs/providers/selectel/d/domains_rrset_v2.html">selectel_domains_rrset_v2</a>
-            </li>
             <li<%= sidebar_current("docs-selectel-datasource-dbaas-datastore-type-v1") %>>
               <a href="/docs/providers/selectel/d/dbaas_datastore_type_v1.html">selectel_dbaas_datastore_type_v1</a>
             </li>
